% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_coops_stations.R
\name{list_port_codes}
\alias{list_port_codes}
\title{Retrieves Unique Port Codes and Port Name}
\usage{
list_port_codes()
}
\value{
A data frame
}
\description{
Provided as a convince function to users it returns all the unique
combinations of port code and port name from \code{\link{query_port_stations()}}
}
\details{
This function produces a data frame with two columns. One column is the two
character code for a port. The second column contains the name of the port
that indicated by the code.
}
\seealso{
\code{\link{query_port_stations}} \code{vignette('stations', package='noaaoceans')}
}
