% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaaoceans.R
\docType{package}
\name{noaaoceans}
\alias{noaaoceans}
\title{noaaoceans: A package for collecting oceans and weather data from NOAA.}
\description{
The noaaoceans package provides several functions to access NOAA APIs.  It
includes functions to access the
\href{https://api.tidesandcurrents.noaa.gov/api/prod/}{CO-OPS API} and metadata for
each of the tide sensor stations.
}
