% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAATides.R
\docType{package}
\name{NOAATides}
\alias{NOAATides}
\alias{NOAATides-package}
\title{NOAATides: A package for collecting tides data from NOAA.}
\description{
The NOAATides package provides several functions to access NOAA APIs.  It
includes funtions to access the
\href{https://tidesandcurrents.noaa.gov/api/}{CO-OPS API} and metadata for
each of the tide sensor stations.
}
