% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_derived_products.R
\name{query_derived_products}
\alias{query_derived_products}
\title{Query Data From NOAA's CO-OPS Derived Product API}
\usage{
query_derived_products(station_id = NULL, product_name = NULL,
  year = NULL, affil = NULL, units = "english")
}
\arguments{
\item{station_id}{an optional string that provides the a 7 character station
id. If omitted the derived product API returns data for all stations.}

\item{product_name}{a string providing the name of the derived data product.
Derived products include Top Ten Water Levels (\code{'toptenwaterlevels'}),
Annual Flood Days (\code{'annualflooddays'}), Extreme Water Levels
(\code{'extremewaterlevels'}) and Sea Level Trends (\code{'sealeveltrends'}).}

\item{year}{an optional string used to limit the results from the annual
flood days endpoint to the indicated year. The argument is ignored if used
with other data products.}

\item{affil}{an argument used to limit the results to U.S. (\code{'US'}) or
Global stations (\code{'Global'}). The argument is ignored if used with
other data products.}

\item{units}{a character string specifying if the data should be returned
using metric or English units. Defaults to \code{'english'}}
}
\value{
a data frame
}
\description{
Provides access to data available from NOAA's CO-OPS \href{https://tidesandcurrents.noaa.gov/dpapi/latest/#intro}{Derived Product API}. Four derived data products are available through the API: 1) Top Ten Water Levels, 2) Annual Flood Days, 3) Extreme Water Levels and 4) Sea Level Trends.  More detail about each data product is available with the \href{https://tidesandcurrents.noaa.gov/dpapi/latest/#intro}{API's documentation}
}
