% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_coops_stations.R
\name{query_port_codes}
\alias{query_port_codes}
\title{Query NOAA PORTS Stations}
\usage{
query_port_codes(port_code = NULL)
}
\arguments{
\item{port_code}{an optional two letter string signifying a collection of
port stations. If it is not provided all port code stations are returned.}
}
\value{
A data frame
}
\description{
NOAA's PORTS stations are a colleciton of integrated sensor stations that are
concentrated in seaports. These stations are intended to commerial vessels
with real time environmental conditions. A list of these stations and
data about the station's location and sensors.
}
\details{
This API is not a documented part of CO-OPS web services. It was discovered
as a part of the documentation
\href{https://api.tidesandcurrents.noaa.gov/mdapi/prod/#intro}{CO-OPS Metadata API}.
}
\examples{
\donttest{
port_code <- query_port_codes()
north_chesapeak <- query_port('cn')
}
}
